Second commit

+++ src/main/java/ru/cloudproject/cloud/cloudtest/controllers/FileController.java
@@ -14,9 +14,12 @@ import ru.cloudproject.cloud.cloudtest.entities.File;
 import ru.cloudproject.cloud.cloudtest.entities.User;
 import ru.cloudproject.cloud.cloudtest.services.FileService;
 import ru.cloudproject.cloud.cloudtest.services.UserService;
+import ru.cloudproject.cloud.cloudtest.services.UserTokenService;
 import ru.cloudproject.cloud.cloudtest.utils.FilenameNotFoundException;
+import ru.cloudproject.cloud.cloudtest.utils.JwtUtil;
 import ru.cloudproject.cloud.cloudtest.utils.UnauthorizedExeption;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.constraints.NotNull;
 import java.io.IOException;
 import java.security.Principal;
@@ -27,49 +30,71 @@ import java.util.*;
 public class FileController {
     private final FileService fileService;
     private final UserService userService;
+    private final UserTokenService userTokenService;
+    private final JwtUtil jwtUtil;
     @Autowired
-    public FileController(FileService fileService, UserService userService) {
+    public FileController(FileService fileService, UserService userService,UserTokenService userTokenService,JwtUtil jwtUtil) {
         this.fileService = fileService;
         this.userService = userService;
+        this.userTokenService = userTokenService;
+        this.jwtUtil = jwtUtil;
     }
 
 
 
 
     @GetMapping("/list")
-    public ResponseEntity<List<FileNameAndSizeDTO>> getAllFiles(Principal principal) {
+    public ResponseEntity<List<FileNameAndSizeDTO>> getAllFiles(@RequestHeader(name = "Authorization") String token,Principal principal) {
+        String email = principal.getName();
+        String normalToken = token.substring(7);
+        if (token == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
-            String email = principal.getName();
+        if (!email.equals(jwtUtil.validateTokenAndRetrieveSubject(normalToken))) {
 
-            List<FileNameAndSizeDTO> files = fileService.getFileList(email);
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        List<FileNameAndSizeDTO> files = fileService.getFileList(email);
 
 
 
-            return new ResponseEntity<>(files,HttpStatus.OK);
+        return new ResponseEntity<>(files,HttpStatus.OK);
 
     }
 
     @PostMapping("/file")
-    public ResponseEntity<?> uploadFile(@RequestParam("file") MultipartFile file) throws IOException {
-        try {
-            fileService.saveFile(file);
-            return ResponseEntity.ok("File uploaded successfully");
+    public ResponseEntity<?> uploadFile(@RequestParam("file") MultipartFile file, @RequestHeader(name = "Authorization") String token, Principal principal) throws IOException {
+        String email = principal.getName();
+        String normalToken = token.substring(7);
 
-        } catch (UnauthorizedExeption e){
+        if (!email.equals(jwtUtil.validateTokenAndRetrieveSubject(normalToken))) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                     .contentType(MediaType.APPLICATION_JSON)
                     .body(new Error("Unauthorized error"));
         }
-        catch (Exception e) {
+
+        try {
+            fileService.saveFile(file);
+            return ResponseEntity.ok("File uploaded successfully");
+        } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .contentType(MediaType.APPLICATION_JSON)
                     .body(new Error("Error upload file"));
         }
-
     }
 
     @DeleteMapping("/file")
-    public ResponseEntity<?> deleteFile(@RequestParam String filename) {
+    public ResponseEntity<?> deleteFile(@RequestParam String filename, @RequestHeader(name = "Authorization") String token, Principal principal) {
+        String email = principal.getName();
+        String normalToken = token.substring(7);
+
+        if (!email.equals(jwtUtil.validateTokenAndRetrieveSubject(normalToken))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .body("Unauthorized Error");
+        }
+
         try {
             fileService.deleteFileByName(filename);
             return ResponseEntity.ok("Success deleted");
@@ -77,11 +102,6 @@ public class FileController {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .contentType(MediaType.APPLICATION_JSON)
                     .body("Error input data");
-
-        } catch (UnauthorizedExeption e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .body("Unauthorized Error");
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .contentType(MediaType.APPLICATION_JSON)
@@ -89,7 +109,16 @@ public class FileController {
         }
     }
     @GetMapping("/file")
-    public ResponseEntity<?> downloadFile(@RequestParam String filename) {
+    public ResponseEntity<?> downloadFile(@RequestParam String filename, @RequestHeader(name = "Authorization") String token, Principal principal) {
+        String email = principal.getName();
+        String normalToken = token.substring(7);
+
+        if (!email.equals(jwtUtil.validateTokenAndRetrieveSubject(normalToken))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .body(new Error("Unauthorized error"));
+        }
+
         try {
             byte[] fileBytes = fileService.downloadFile(filename);
             if (fileBytes == null) {
@@ -102,10 +131,6 @@ public class FileController {
             return ResponseEntity.ok()
                     .contentType(MediaType.APPLICATION_OCTET_STREAM)
                     .body(resource);
-        } catch (UnauthorizedExeption e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .body(new Error("Unauthorized error"));
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .contentType(MediaType.APPLICATION_JSON)
@@ -117,7 +142,18 @@ public class FileController {
 
     @PutMapping("/file")
     public ResponseEntity<?> editFileName(@RequestParam(name = "filename") String filename,
-                                          @RequestBody Map<String, String> requestBody) {
+                                          @RequestBody Map<String, String> requestBody,
+                                          @RequestHeader(name = "Authorization") String token,
+                                          Principal principal) {
+        String email = principal.getName();
+        String normalToken = token.substring(7);
+
+        if (!email.equals(jwtUtil.validateTokenAndRetrieveSubject(normalToken))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .body("Unauthorized Error");
+        }
+
         try {
             String newName = requestBody.get("name");
             if (newName == null || newName.isEmpty()) {
@@ -130,11 +166,6 @@ public class FileController {
                         .contentType(MediaType.APPLICATION_JSON)
                         .body("Success upload");
             }
-
-        } catch (UnauthorizedExeption e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .body("Unauthorized Error");
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .contentType(MediaType.APPLICATION_JSON)
